+++ R/convert_cols_module.R
@@ -44,15 +44,15 @@ convert_cols_ui <- function(id) {
 }
 
 # server ----------------------------------------------------------------------
-convert_cols_server <- function(id, input_df) {
+convert_cols_server <- function(id) {
   moduleServer(id, function(input, output, session) {
 	  ns <- session$ns
 
-    df_names <- reactive(input_df() |> names())
+    df_names <- reactive(session$userData$df$act |> names())
 
     df <- reactiveValues()
     observe({
-      df$df_active <- input_df()
+      df$df_active <- session$userData$df$act
     })
 
     output$ui_var_sel <- renderUI(
@@ -145,6 +145,11 @@ convert_cols_server <- function(id, input_df) {
       }
     }) |> bindEvent(input$btn_apply)
 
-    return(list(df_convert_cols = reactive(df$df_active)))
+    # update active dataset ---------------------------------------------------
+    observe({
+      req(df$df_active)
+      session$userData$df$act <- df$df_active
+    })
+
   })
 }