+++ R/correlation_module.R
@@ -65,15 +65,17 @@ correlation_ui <- function(id) {
 }
 
 # server ----------------------------------------------------------------------
-correlation_server <- function(id, df, df_metadata, output_report) {
+correlation_server <- function(id) {
   moduleServer(id, function(input, output, session) {
 	  ns <- session$ns
+
+	  df <- reactive(session$userData$df$act)
 
     # outupt objects ----------------------------------------------------------
 	  output_list <- reactiveValues(elements = NULL)
 
 	  observe({
-	    output_list$elements <- output_report()
+	    output_list$elements <- session$userData$out$elements
 	  })
 
     cor_test <- reactiveValues(results = NULL)
@@ -83,7 +85,7 @@ correlation_server <- function(id, df, df_metadata, output_report) {
     var_analysis <- reactive({
       df_names <- df_active() |> names()
 
-      var_analysis <- df_metadata() |> filter(perc_nas != 1) |> pull(var)
+      var_analysis <- session$userData$df$act_meta() |> filter(perc_nas != 1) |> pull(var)
 
       df_names[df_names %in% var_analysis]
       })
@@ -242,8 +244,10 @@ correlation_server <- function(id, df, df_metadata, output_report) {
       ))
     }) |> bindEvent(input$btn_help_cor)
 
-    # return values -----------------------------------------------------------
-    return(list(output_file = reactive(output_list$elements)))
+    # update output -----------------------------------------------------------
+    observe({
+      session$userData$out$elements <- output_list$elements
+    })
 
   })
 }