+++ R/order_cols_module.R
@@ -22,16 +22,16 @@ order_cols_ui <- function(id) {
 }
 
 # server ----------------------------------------------------------------------
-order_cols_server <- function(id, input_df) {
+order_cols_server <- function(id) {
   moduleServer(id, function(input, output, session) {
 	  ns <- session$ns
 
     # Reactive to get column names
-    df_names <- reactive(input_df() |> names())
+    df_names <- reactive(session$userData$df$act |> names())
 
     df <- reactiveValues()
     observe({
-      df$df_active <- input_df()
+      df$df_active <- session$userData$df$act
     })
 
     output$ui_var_cols <- renderUI(
@@ -74,6 +74,10 @@ order_cols_server <- function(id, input_df) {
       }
     }) |> bindEvent(input$btn_order_cols)
 
-    return(list(df_order_cols = reactive(df$df_active)))
+    # update active dataset ---------------------------------------------------
+    observe({
+      req(df$df_active)
+      session$userData$df$act <- df$df_active
+    })
   })
 }