+++ R/select_cols_module.R
@@ -12,20 +12,19 @@ select_cols_ui <- function(id) {
     ),
     card_footer(btn_task(ns('btn_sel'), 'Apply selection', icon('check')))
   )
-
 }
 
 # server ----------------------------------------------------------------------
-select_cols_server <- function(id, input_df) {
+select_cols_server <- function(id) {
   moduleServer(id, function(input, output, session) {
 	  ns <- session$ns
 
     # Reactive to get column names
-    df_names <- reactive(input_df() |> names())
+    df_names <- reactive(session$userData$df$act |> names())
 
     df <- reactiveValues()
     observe({
-      df$df_active <- input_df()
+      df$df_active <- session$userData$df$act
     })
 
     output$ui_var_sel <- renderUI(
@@ -62,6 +61,11 @@ select_cols_server <- function(id, input_df) {
       }
     }) |> bindEvent(input$btn_sel)
 
-    return(list(df_sel_cols = reactive(df$df_active)))
+    # update active dataset ---------------------------------------------------
+    observe({
+      req(df$df_active)
+      session$userData$df$act <- df$df_active
+    })
+
   })
 }