+++ tests/testthat/test-data_overview_module.R
@@ -1,9 +1,8 @@
 # tests/testthat/test-data_overview_module.R
 
-
 # test validate number of rows ------------------------------------------------
 test_that("Test sample rows", {
-  testServer(data_overview_server, args = list(df = reactive(iris)), {
+  testServer(data_overview_server, {
     session$setInputs(size_sample = 0)
     expect_error(idx(), "Number of rows must be > 0")
   })
@@ -11,8 +10,9 @@ test_that("Test sample rows", {
 
 # test first rows -------------------------------------------------------------
 test_that("Test first rows", {
-  testServer(data_overview_server, args = list(df = reactive(iris)), {
+  testServer(data_overview_server, {
 
+    session$userData$df <- reactiveValues(act = iris)
     # Simulate inputs
     session$setInputs(size_sample = 5)
     session$setInputs(radio_sample = "first")
@@ -24,7 +24,10 @@ test_that("Test first rows", {
 
 # test sample of rows ---------------------------------------------------------
 test_that("Test sample rows", {
-  testServer(data_overview_server, args = list(df = reactive(iris)), {
+  testServer(data_overview_server, {
+
+    session$userData$df <- reactiveValues(act = iris)
+
     # Simulate inputs
     session$setInputs(size_sample = 5)
     session$setInputs(radio_sample = "sample")